==> regulon/03_heatmap_regulons.Rmd <==
---
title: "Regulons and expression heatmaps"
author: "Sofia Salazar"
date: "2024-08-16"
output: html_document
---

Libraries

```{r}
library(tidyverse)
library(ComplexHeatmap)
library(circlize)
library(DESeq2)
library(ggplot2)
```


```{r}
name_short<-function(table){
  table$ID <- gsub("\\..*","", table$ID)
  return(table)
}
```

Data

```{r}
workdir <- "/mnt/Citosina/amedina/ssalazar/meta/combined/"
regulon_dir <- glue::glue("{workdir}/regulons")
results_dir <- glue::glue("{regulon_dir}/results")
DEregulon_targets <- read.csv(glue::glue("{results_dir}/DEregulon_targets.csv"), row.names = "X")
load("/mnt/Citosina/amedina/backup/lupus/sofi/vsd2.RData")
load("/mnt/Citosina/amedina/ssalazar/meta/combined/namedDGElist.RData")

all_data <-  read.csv(glue::glue("{workdir}/all_data.csv"), header = T)
  
ordered_samples <- all_data[order(all_data$DISEASE),] # reorder all_data



norm_counts <- as.data.frame(assay(vsd2))
norm_counts <- tibble::rownames_to_column(norm_counts, "ID")
norm_counts_id <- name_short(norm_counts)
norm_counts_name <- inner_join(norm_counts_id, df_names, by = 'ID')

all(df_names$gene_name == norm_counts_name$gene_name) # TRUE
norm_counts_name<- norm_counts_name[,-c(1,320:327)]
rownames(norm_counts_name) <- df_names$gene_name

counts <- as.matrix(norm_counts_name)
zscore <- t(scale(t(counts)))

```


Get counts matrix, some rows are repeated because a gene is in more than one regulon

```{r}
filtered_DEregulon_targets <- DEregulon_targets %>%
  filter(geneName %in% rownames(zscore))

dim(filtered_DEregulon_targets)

ordered_counts <- zscore[filtered_DEregulon_targets$geneName, , drop = FALSE]
all(rownames(ordered_counts) == filtered_DEregulon_targets$geneName) # TRUE
ordered_counts <- ordered_counts[,ordered_samples$samples]
all(colnames(ordered_counts) == ordered_samples$samples)

```


Terms of each regulon

```{r}

parentTerms_up <- read.csv(glue::glue("{results_dir}/parentTerms_up.csv"),
                             header = T, row.names = "X")
parentTerms_down <- read.csv(glue::glue("{results_dir}/parentTerms_down.csv"),
                             header = T, row.names = "X")

parentTerms <- rbind(parentTerms_down, parentTerms_up)
reg_df <- filtered_DEregulon_targets %>%
  left_join(parentTerms %>% select(regulon, Small), by = "regulon") %>%
  arrange(-log2FC)


term_col <- c("Cellular transport" = "#d6b3f2", "Response to stimulus" = "#f0b800", 
  "Immune cell activation" = "#0f5203", "Metabolic process" = "#a3a5cf", "Developmental process"  = "#de049d", "Cell cycle"  = "#08a3a6", "Nucleic acid processing" = "#b3de7e", "Immune system process" = "#ff7300", "Neural process" = "#f54545", "TF activity"="#005cb8", "Dendritic cell differentiation" = "#6b08c2")

```

Heatmap

```{r}
unique(reg_df$log2FC)
ordered_genes <- reg_df[order(-reg_df$log2FC),]


ordered_counts <- ordered_counts[ordered_genes$geneName,]

all(rownames(ordered_counts) == reg_df$geneName)
ha <- HeatmapAnnotation(Samples = ordered_samples$DISEASE,
                        col = list(Samples = c('CONTROL' = '#96d4ccff', 'SLE' = '#b493b4ff')))

l2_val <- as.matrix(reg_df$log2FC)
col_lfc <- colorRamp2(c(min(l2_val), 0, 0.1,  max(l2_val)), c("blue", "white", "#ffb5b0","red"))

lfc_anno = rowAnnotation("log2FC" = reg_df$log2FC, col = list("log2FC" = col_lfc))

term_anno <- rowAnnotation("Regulon associated process" = reg_df$Small,
                           col = list("Regulon associated process" = term_col), 
                           show_annotation_name = F)
sample_split = data.frame("Sample" = ordered_samples$DISEASE)


color_fun <- colorRamp2(c(min(ordered_counts), 0, 2, max(ordered_counts)), c('blue', 'white', 'red','darkred'))

direction_split = cut(l2_val, breaks = c(-Inf, 0, Inf), labels = c("Downregulated", "Upregulated"), right = FALSE)
h <- Heatmap(ordered_counts, name = "Z-score", col = color_fun, cluster_rows = F, cluster_columns = F, show_row_names = F, show_column_names = F, 
             top_annotation = ha, column_split = sample_split, 
             row_split = factor(direction_split, levels = c("Upregulated", "Downregulated")),
             right_annotation =  c(lfc_anno, term_anno))

png(filename = glue::glue("{results_dir}/heatmap_targetsRegulons.png"), height = 20, width = 20, units = "cm", res = 300)

  draw(h)

dev.off()

# unique_small <- unique(reg_df$Small)
# row_indices <- sapply(unique_small, function(x) which(reg_df$Small == x))
# 
# # Draw rectangles around the blocks of rows
# for(i in seq_along(row_indices)) {
#   decorate_heatmap_body("Z-score", {
#     if(length(row_indices[[i]]) > 1) {
#       row_range <- range(row_indices[[i]])
#       grid.rect(y = unit(row_range[1] - 0.5, "native"), 
#                 height = unit(diff(row_range) + 1, "native"), 
#                 gp = gpar(col = "black", fill = NA, lwd = 2))
#     }
#   })
# }
# 
# png(filename = glue::glue("{results_dir}/heatmap_targetsRegulons.png"), height = 20, width = 20, units = "cm", res = 300)
# 
#   draw(h)
# 
# dev.off()


```

## Regulons associated to immune functions

```{r}
library(tidyverse)

functions_interest <- c("Immune system process", "Immune cell activation", "Dendritic cell differentiation")

immune_regulons <- parentTerms %>%
  filter(Small %in% functions_interest)%>%
  inner_join(filtered_DEregulon_targets, by = "regulon")%>%
  arrange(-log2FC)


immune_regulons$regulon <- factor(immune_regulons$regulon, levels = unique(immune_regulons$regulon))

unique(immune_regulons$regulon)
# [1] STAT1  ETV6   BCL3   IRF4   ZIK1   BCL11A KLF13  TBX21  TCF7 

dim(immune_regulons) # [1] 4980    7
small_counts <- zscore[immune_regulons$geneName, , drop = FALSE]

small_counts <- small_counts[,ordered_samples$samples]
```


```{r}
mat2 <- as.matrix(small_counts)

ha <- HeatmapAnnotation(Samples = ordered_samples$DISEASE,
                        col = list(Samples = c('CONTROL' = '#96d4ccff', 'SLE' = '#b493b4ff')))

col_lfc <- colorRamp2(c(min(immune_regulons$log2FC), 0, max(immune_regulons$log2FC)), c("blue", "white", "red"))

lfc_anno <- rowAnnotation("log2FC" = immune_regulons$log2FC, col = list("log2FC" = col_lfc))
term_col <- c("Immune cell activation" = "#0f5203", "Immune system process" = "#ff7300", "Dendritic cell differentiation" = "#6b08c2")

term_anno <- rowAnnotation("Regulon associated process" = immune_regulons$Small,
                           col = list("Regulon associated process" = term_col),
                           show_annotation_name = F)

regulon_split <- data.frame("Regulon" = immune_regulons$regulon)
regulon_split$Regulon <- factor(regulon_split$Regulon, levels = unique(immune_regulons$regulon))


all(unique(immune_regulons$regulon) == unique(regulon_split$Regulon))
sample_split <- data.frame("Sample" = ordered_samples$DISEASE)
color_fun <- colorRamp2(c(min(mat2), 0, 2, max(mat2)), c('blue', 'white', 'red', 'darkred'))

# Generate the heatmap
h <- Heatmap(mat2, name = "Z-score", col = color_fun, cluster_rows = F, cluster_columns = F, 
             show_row_names = F, show_column_names = F, top_annotation = ha, 
             column_split = sample_split, row_split = regulon_split, 
             right_annotation = c(lfc_anno, term_anno), row_title_rot = 0, show_row_dend = F)

png(filename = glue::glue("{results_dir}/heatmapSmall_targetsRegulons.png"), height = 25, width = 30, units = "cm", res = 300)

  draw(h)

dev.off()
```

# Only TFs

```{r}
tfs_df <- unique(reg_df %>% select(-c(geneName)))
tfs_df <- tfs_df %>%
  group_by(direction) %>%
  arrange(desc(direction),-log2FC) %>%
  ungroup()

ordered_tf <- tfs_df[order(-tfs_df$log2FC),]
tf_counts <- zscore[ordered_tf$regulon, , drop = FALSE]

all(rownames(tf_counts) == tfs_df$regulon)
tf_counts <- tf_counts[,ordered_samples$samples]


mat3 <- as.matrix(tf_counts)
ha <- HeatmapAnnotation(Samples = ordered_samples$DISEASE,
                        col = list(Samples = c('CONTROL' = '#96d4ccff', 'SLE' = '#b493b4ff')))

col_lfc <- colorRamp2(c(min(tfs_df$log2FC), 0, max(tfs_df$log2FC)), c("blue", "white", "red"))

lfc_anno = rowAnnotation("log2FC" = tfs_df$log2FC, col = list("log2FC" = col_lfc))
unique(tfs_df$Small)
term_col <- c("Cellular transport" = "#d6b3f2", "Response to stimulus" = "#f0b800", 
  "Immune cell activation" = "#0f5203", "Metabolic process" = "#a3a5cf", "Developmental process"  = "#de049d", "Cell cycle"  = "#08a3a6", "Nucleic acid processing" = "#b3de7e", "Immune system process" = "#ff7300", "Neural process" = "#f54545", "TF activity"="#005cb8", "Dendritic cell differentiation" = "#6b08c2")

term_anno <- rowAnnotation("Regulon associated process" = tfs_df$Small,
                           col = list("Regulon associated process" = term_col), 
                           show_annotation_name = F)

l2_val <- as.matrix(tfs_df$log2FC)
direction_split = cut(l2_val, breaks = c(-Inf, 0, Inf), labels = c("Downregulated", "Upregulated"), right = FALSE)
sample_split = data.frame("Sample" = ordered_samples$DISEASE)
color_fun <- colorRamp2(c(min(mat3), 0, 2, max(mat3)), c('blue', 'white', 'red','darkred'))

h <- Heatmap(mat3, name = "Z-score", col = color_fun, cluster_rows = F, cluster_columns = F, show_row_names = T, show_column_names = F, 
             top_annotation = ha, column_split = sample_split,
             row_split = factor(direction_split, levels = c("Upregulated", "Downregulated")),
              right_annotation =  c(lfc_anno,term_anno), row_title_rot = 0, show_row_dend = F)

png(filename = glue::glue("{results_dir}/heatmapTFs_Regulons.png"), height = 36, width = 30, units = "cm", res = 300)
draw(h)
dev.off()
```


```{r}
sessionInfo()

# R version 4.0.2 (2020-06-22)
# Platform: x86_64-pc-linux-gnu (64-bit)
# Running under: CentOS Linux 7 (Core)
# 
# Matrix products: default
# BLAS:   /cm/shared/apps/r/4.0.2-studio/lib64/R/lib/libRblas.so
# LAPACK: /cm/shared/apps/r/4.0.2-studio/lib64/R/lib/libRlapack.so
# 
# locale:
#  [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C              
#  [3] LC_TIME=en_US.UTF-8        LC_COLLATE=en_US.UTF-8    
#  [5] LC_MONETARY=en_US.UTF-8    LC_MESSAGES=en_US.UTF-8   
#  [7] LC_PAPER=en_US.UTF-8       LC_NAME=C                 
#  [9] LC_ADDRESS=C               LC_TELEPHONE=C            
# [11] LC_MEASUREMENT=en_US.UTF-8 LC_IDENTIFICATION=C       
# 
# attached base packages:
#  [1] parallel  stats4    grid      stats     graphics  grDevices utils    
#  [8] datasets  methods   base     
# 
# other attached packages:
#  [1] DESeq2_1.30.1               SummarizedExperiment_1.20.0
#  [3] Biobase_2.50.0              MatrixGenerics_1.2.1       
#  [5] matrixStats_1.0.0           GenomicRanges_1.42.0       
#  [7] GenomeInfoDb_1.30.1         IRanges_2.24.1             
#  [9] S4Vectors_0.28.1            BiocGenerics_0.38.0        
# [11] circlize_0.4.14             ComplexHeatmap_2.9.3       
# [13] forcats_0.5.1               stringr_1.4.0              
# [15] dplyr_1.0.10                purrr_0.3.4                
# [17] readr_1.4.0                 tidyr_1.2.1                
# [19] tibble_3.1.3                ggplot2_3.5.1              
# [21] tidyverse_1.3.0            
# 
# loaded via a namespace (and not attached):
#  [1] bitops_1.0-7           fs_1.5.0               bit64_4.0.5           
#  [4] lubridate_1.7.9.2      doParallel_1.0.17      RColorBrewer_1.1-2    
#  [7] httr_1.4.2             tools_4.0.2            backports_1.2.1       
# [10] utf8_1.2.2             R6_2.5.0               DBI_1.1.1             
# [13] colorspace_2.0-2       GetoptLong_1.0.5       withr_3.0.0           
# [16] tidyselect_1.2.0       bit_4.0.4              compiler_4.0.2        
# [19] cli_3.6.0              rvest_0.3.6            xml2_1.3.2            
# [22] DelayedArray_0.16.3    scales_1.3.0           genefilter_1.72.1     
# [25] digest_0.6.27          XVector_0.30.0         pkgconfig_2.0.3       
# [28] fastmap_1.1.0          dbplyr_2.2.1           rlang_1.1.3           
# [31] GlobalOptions_0.1.2    readxl_1.3.1           rstudioapi_0.13       
# [34] RSQLite_2.2.7          shape_1.4.6            generics_0.1.3        
# [37] jsonlite_1.7.2         BiocParallel_1.24.1    RCurl_1.98-1.3        
# [40] magrittr_2.0.1         GenomeInfoDbData_1.2.4 Matrix_1.6-5          
# [43] Rcpp_1.0.7             munsell_0.5.0          fansi_0.5.0           
# [46] lifecycle_1.0.3        stringi_1.6.2          zlibbioc_1.36.0       
# [49] blob_1.2.1             crayon_1.4.1           lattice_0.20-41       
# [52] splines_4.0.2          haven_2.3.1            annotate_1.68.0       
# [55] hms_1.1.0              locfit_1.5-9.4         pillar_1.6.2          
# [58] rjson_0.2.20           geneplotter_1.68.0     codetools_0.2-18      
# [61] reprex_1.0.0           XML_3.99-0.6           glue_1.4.2            
# [64] modelr_0.1.8           png_0.1-7              vctrs_0.6.5           
# [67] foreach_1.5.2          cellranger_1.1.0       gtable_0.3.0          
# [70] clue_0.3-59            assertthat_0.2.1       cachem_1.0.5          
# [73] xtable_1.8-4           broom_0.7.9            survival_3.5-5        
# [76] iterators_1.0.13       memoise_2.0.0          AnnotationDbi_1.52.0  
# [79] cluster_2.1.0          ellipsis_0.3.2 
```
